Mark create-acme-provisioner as a PostSync hook
We want the create-acme-provisioner job to be cleaned up after
successful completion. Previously we tried using
ttlSecondsAfterFinished on the job, which causes it to be clean up by
Kubernetes, but in this case ArgoCD automatically re-creates the Job,
causing it to be scheduled forever.

By marking it as an ArgoCD PostSync hook and setting a
hook-delete-policy, we can have ArgoCD take care of cleaning up
instead.